Create Leetcode 143. Reorder List.java
You are given the head of a singly linked-list. The list can be represented as:

L0 → L1 → … → Ln - 1 → Ln
Reorder the list to be on the following form:

L0 → Ln → L1 → Ln - 1 → L2 → Ln - 2 → …
You may not modify the values in the list's nodes. Only nodes themselves may be changed.

Solution- 1st. Find middle of the list
2nd. Reverse the 2nd half of list
3rd. Now arrange the list